LayoutTests/imported/w3c/web-platform-tests/loading/early-hints/modulepreload-as-worker.h2.window-expected.txt


FAIL Modulepreload should load with as="worker" from same-origin url assert_equals: expected true but got false
